#32 - Moving objects folder out of scenes folder
