+++ client/src/app/scenograph/objects/bot.js
