+++ client/src/app/scenograph/objects/cargo_ships.js
