+++ client/src/app/scenograph/objects/extractors.js
