+++ client/src/app/scenograph/objects/office_room.js
