+++ client/src/app/scenograph/objects/platform.js
