+++ client/src/app/scenograph/objects/player.js
