+++ client/src/app/scenograph/objects/sky.js
