+++ client/src/app/scenograph/objects/sky2.js
